+++ src/resources/views/inc/renderButton.blade.php
@@ -1,0 +1,5 @@
+@if ($button->type == 'model_function')
+	{!! $entry->{$function_name}(); !!}
+@else
+	@include($button->content)
+@endif